requirements.txt
# Core Dependencies
flask==3.0.0
werkzeug==3.0.1
celery==5.3.6
redis==5.0.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
jinja2==3.1.2
watchdog==3.0.0
feedparser==6.0.10

# Computer Vision & Face Recognition
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2

# Audio Processing & Speaker Diarization
pyannote.audio==3.1.1
torch==2.1.0
torchaudio==2.1.0

# Data Processing
numpy==1.24.3

# System Integration
python-dotenv==1.0.0

# Optional: Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
